refactor(mattermost): move configuration types into mattermost package
